Address review comments: fix error messages, improve docs, move XORShiftRandom to lib
- Fix error message consistency for length and seed arguments
- Change 'partition' to 'partition ID' in documentation
- Improve wording for partition config description
- Move XORShiftRandom.h to velox/functions/lib for common use
